+++ app/src/main/java/bitcamp/carrot_thunder/dto/PostRequestDto.java
@@ -1,15 +1,21 @@
 package bitcamp.carrot_thunder.dto;
 
-import bitcamp.carrot_thunder.post.model.vo.DealingType;
-import bitcamp.carrot_thunder.post.model.vo.ItemCategory;
+import bitcamp.carrot_thunder.post.model.vo.AttachedFile;
 import lombok.Getter;
+
+import java.util.List;
 
 @Getter
 public class PostRequestDto {
 
-    private String title;
-    private String content;
-    private int price;
-    private DealingType dealingType;
-    private ItemCategory itemCategory;
+  private String title;
+  private String content;
+  private String itemCategory;
+  private String dealingType;
+  private String itemStatus;
+  private String address;
+  private int price;
+  private List<AttachedFile> attachedFilesPaths;
+
+//이걸로게시글작성
 }
