+++ app/src/main/java/bitcamp/carrot_thunder/dto/PostUpdateRequestDto.java
@@ -1,4 +1,16 @@
 package bitcamp.carrot_thunder.dto;
 
+import bitcamp.carrot_thunder.post.model.vo.AttachedFile;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
 public class PostUpdateRequestDto {
+  private String title;
+  private String content;
+  private int price;
+  private String address;
+  private List<AttachedFile> attachedFilesPaths;
+
 }
