--- app/src/main/java/bitcamp/carrot_thunder/post/dto/PostRequestDto.java
@@ -1,21 +1,0 @@
-package bitcamp.carrot_thunder.post.dto;
-
-import bitcamp.carrot_thunder.post.model.vo.AttachedFile;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class PostRequestDto {
-
-  private String title;
-  private String content;
-  private String itemCategory;
-  private String dealingType;
-  private String itemStatus;
-  private String address;
-  private int price;
-  private List<AttachedFile> attachedFilesPaths;
-
-//이걸로게시글작성
-}
