+++ app/src/main/java/bitcamp/carrot_thunder/post/model/vo/Post.java
@@ -1,10 +1,6 @@
 package bitcamp.carrot_thunder.post.model.vo;
 
-import bitcamp.carrot_thunder.post.dto.PostUpdateRequestDto;
-import bitcamp.carrot_thunder.post.model.vo.AttachedFile;
-import bitcamp.carrot_thunder.post.model.vo.DealingType;
-import bitcamp.carrot_thunder.post.model.vo.ItemCategory;
-import bitcamp.carrot_thunder.post.model.vo.ItemStatus;
+import bitcamp.carrot_thunder.dto.PostUpdateRequestDto;
 //import bitcamp.carrot_thunder.user.model.vo.User;
 import bitcamp.carrot_thunder.user.model.vo.User;
 import lombok.AllArgsConstructor;
